Altered File Hierarchy rendering seperate FileContent objects redundant, found out str.isidentifier exists
