--- webgoose/struct/component/__init__.py
@@ -1,3 +1,0 @@
-# webgoose.struct.component
-
-from    .component      import      Component
